plots the nodal plots can only do it if ONE date is selected, I will have to modify it to either do multiple dates or just deal with multiple dates and plot the one date. #162
