+++ layout/NodalAnalysis.html
@@ -1,0 +1,57 @@
+<style>
+  .pad-top{
+    padding-top: 10px;
+  }
+
+  .pad-bottom {
+    padding-bottom: 10px;
+  }
+
+  .th-lg {
+    font-size: 15pt;
+  }
+
+  .th-md {
+    font-size: 9pt;
+  }
+</style>
+
+<link href="/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+
+<div class="container">
+  <div class="row">
+    <div class="col-2 Field-Well-Name-List">
+      <pre style="font-family: Verdana; font-weight: bold; font-size: 13pt;">Select satellite(s):    Select wells(s):</pre>
+      <SpotfireControl id="736a5598c390417d96503e85e126cc7a" />
+      <SpotfireControl id="c4284a80e58c49e6a94c9861d8c22adf" />
+    </div>
+
+    <div class="col-2 Reset-Filters pad-bottom">
+      <SpotfireControl id="8b0df0029eeb42ba9dbad3ff2c707c82" />
+    </div>
+
+    <div class="col-2 ProVe-Analysis pad-top pad-bottom">
+      <SpotfireControl id="ac56f7b6929843e99262099fdbe25fbb" />
+    </div>
+  </div>
+
+  <div class="row pad-top">
+    <div class="col-2 WEBSITE LINK">
+      <a class="th-md" href="https://joseph-tripodi.github.io/prove-website/documentation.html"> Need help?</a>
+    </div>
+
+    <div class="col-2 Feedback pad-top">
+      <a class="th-md" href="mailto:prove.analysis@gmail.com?Subject=Feedback">Send Feedback</a>
+    </div>
+    <div class="col-2 Error Messages pad-top">
+      <a class="th-md" href="mailto:prove.analysis@gmail.com?Subject=Error%20Messages">Send Error messages</a>
+      <p class="th-md">
+        <b>* Copy the error message from spotfire and paste into email body.</b> <br>
+        <em>** I am aware that it sometimes opens up two email windows, I must apologise<br> for that. but hopefully
+        you will not have to use this very often.</em>
+      </p>
+
+    </div>
+  </div>
+
+</div>
